Create README.md
This is a real-time application that can be used for employees/students/anywhere. For example, if employees use this he can login and logout 
easily with a single click and from admins side, the admin can easily manage all of the user data.He can export user data and use this information 
to generate pay for his employees. If a professor wants to use this, He can calculate his students data and only modification that we will do is to the "pay" column. Instead of 
generating pay we will write the code for handling attendance of students. This application is not only restricted to Students/University/ Employees,
It can be used at Super Markets for calculating sales and for share holders to keep a track of his shares (all real time data can be handled)